Refactor SimpleDialog title handling
Updated the SimpleDialog component to ensure the title is always displayed, defaulting to "Dialog" when no title is provided. This improves accessibility by using "sr-only" for screen readers when the title is absent.

It was showing the following issue when opening the user settings from the dropdown

" DialogContent requires a DialogTitle for the component to be accessible for screen reader users.

If you want to hide the DialogTitle, you can wrap it with our VisuallyHidden component.

For more information, see https://radix-ui.com/primitives/docs/components/dialog"